merge frontend into dev

+++ backend/src/models/User.js
@@ -2,7 +2,9 @@ import { pool } from "../config/db.js";
 
 export class User {
   static async getUsers() {
-    const data = await pool.query(`SELECT * FROM users`);
+    const data = await pool.query(
+      `SELECT user_id, name, surname, patronymic, login, id_role  * FROM users`,
+    );
     return data.rows;
   }
 
@@ -20,7 +22,7 @@ export class User {
         patronymic,
         login,
         password,
-        id_role) VALUES($1, $2, $3, $4, $5, $6) RETURNING *`,
+        id_role) VALUES($1, $2, $3, $4, $5, $6) RETURNING user_id, name, surname, patronymic, login, id_role`,
       [name, surname, patronymic, login, password, id_role],
     );
 
@@ -31,7 +33,7 @@ export class User {
 
   static async deleteUser(id_user) {
     const data = await pool.query(
-      `DELETE FROM users WHERE id_user=$1 RETURNING *`,
+      `DELETE FROM users WHERE id_user=$1 RETURNING user_id, name, surname, patronymic, login, id_role`,
       [id_user],
     );
 
@@ -41,9 +43,10 @@ export class User {
   }
 
   static async getUser(id_user) {
-    const data = await pool.query(`SELECT * FROM users WHERE id_user=$1`, [
-      id_user,
-    ]);
+    const data = await pool.query(
+      `SELECT user_id, name, surname, patronymic, login, id_role FROM users WHERE id_user=$1`,
+      [id_user],
+    );
 
     if (!data.rows.length) throw { status: 404, message: "User Not Found" };
 
@@ -60,11 +63,31 @@ export class User {
     return data.rows[0];
   }
 
-  // static async updateUser({ id_user, params }) {
-  //   const { field, value } = params;
-  //   const data = await pool.query(
-  //     `UPDATE users SET ${field}=$1 WHERE id_user=$2`,
-  //     [value, id_user],
-  //   );
-  // } доделать редактирование юзеров
+  static async updateUser(form) {
+    const updatedData = {
+      name: form.name,
+      surname: form.surname,
+      patronymic: form.patronymic,
+      login: form.login,
+      id_role: form.id_role,
+    };
+
+    const id_user = form.id_user;
+
+    const data = await pool.query(
+      `UPDATE users SET name=$1, surname=$2, patronymic=$3, login=$4, id_role=$5 WHERE id_user=$6 RETURNING id_user, name, surname, patronymic, login, id_role`,
+      [
+        updatedData.name,
+        updatedData.surname,
+        updatedData.patronymic,
+        updatedData.login,
+        updatedData.id_role,
+        id_user,
+      ],
+    );
+
+    if (!data.rows[0]) throw { status: 400, message: "Invalid data" };
+
+    return data.rows[0];
+  }
 }
